compile

+++ proposals/mips/mip-o12/mip-o12.sol
@@ -114,7 +114,7 @@ contract mipo12 is HybridProposal {
             uint256 expectedStartedAt,
             uint256 expectedUpdatedAt,
             uint80 expectedAnsweredInRound
-        ) = wrapper.originalFeed().getRoundData(lastRoundId - 1);
+        ) = wrapper.originalFeed().getRoundData(uint80(lastRoundId - 1));
 
         (
             uint80 actualRoundId,
